+++ src/ts/components/hud/LowerInfoHudFragment.ts
@@ -64,33 +64,58 @@ export class LowerInfoHudFragment implements RenderableComponent {
     // Draw HUD in the lower right corner
     this.hudComponent.hudBitmap.fillStyle = getColorStyle(this.hudComponent.primaryColor, 0.75);
 
+    const offsetY = hudBounds.min.y + tweakParams.lineHeight;
+
+    const hudTextA: string = `Depth: ${hudData.depth.toFixed(1)}m\n`;
+    const hudTextB: string = `Grnd: ${hudData.groundDepth.toFixed(1)}m\n`;
+    const hudTextC: string = `Angle: ${(hudData.shipRotation.upAngle * RAD2DEG).toFixed(1)}`;
+    const hudTextD: string = `Press: ${hudData.pressure.toFixed(1)}bar`;
+    // °F = °C × (9/5) + 32
+    const degFahrenheit = hudData.temperature * (9 / 5) + 32.0;
+    const hudTextE: string = ` Temp: ${hudData.temperature.toFixed(1)}°C / ${degFahrenheit.toFixed(1)}°F`;
+    const textArray = [hudTextA, hudTextB, hudTextC, hudTextD, hudTextE];
+    for (var i = 0; i < textArray.length; i++) {
+      this.hudComponent.hudBitmap.fillText(
+        textArray[i],
+        this.hudComponent.hudCanvas.width - hudBounds.width / 2,
+        offsetY + i * tweakParams.lineHeight
+      );
+    }
+
+    /*
     const hudTextA: string = `Depth: ${hudData.depth.toFixed(1)}m\n`;
     this.hudComponent.hudBitmap.fillText(
       hudTextA,
       this.hudComponent.hudCanvas.width - hudBounds.width / 2,
-      this.hudComponent.hudCanvas.height - hudBounds.height / 2 - tweakParams.lineHeight / 2
+      offsetY - tweakParams.lineHeight / 2
     );
-    const hudTextB: string = `Angle: ${(hudData.shipRotation.upAngle * RAD2DEG).toFixed(1)}`;
+    const hudTextB: string = `Grnd: ${hudData.groundDepth.toFixed(1)}m\n`;
     this.hudComponent.hudBitmap.fillText(
       hudTextB,
       this.hudComponent.hudCanvas.width - hudBounds.width / 2,
-      this.hudComponent.hudCanvas.height - hudBounds.height / 2 + tweakParams.lineHeight / 2
+      offsetY - tweakParams.lineHeight / 2
     );
-    const hudTextC: string = `Press: ${hudData.pressure.toFixed(1)}bar`;
+    const hudTextC: string = `Angle: ${(hudData.shipRotation.upAngle * RAD2DEG).toFixed(1)}`;
     this.hudComponent.hudBitmap.fillText(
       hudTextC,
       this.hudComponent.hudCanvas.width - hudBounds.width / 2,
-      this.hudComponent.hudCanvas.height - hudBounds.height / 2 + tweakParams.lineHeight * 1.5
+      offsetY + tweakParams.lineHeight / 2
     );
-    // °F = °C × (9/5) + 32
-    const degFahrenheit = hudData.temperature * (9 / 5) + 32.0;
-    const hudTextD: string = ` Temp: ${hudData.temperature.toFixed(1)}°C / ${degFahrenheit.toFixed(1)}°F`;
+    const hudTextD: string = `Press: ${hudData.pressure.toFixed(1)}bar`;
     this.hudComponent.hudBitmap.fillText(
       hudTextD,
       this.hudComponent.hudCanvas.width - hudBounds.width / 2,
-      this.hudComponent.hudCanvas.height - hudBounds.height / 2 + tweakParams.lineHeight * 2.5
+      offsetY + tweakParams.lineHeight * 1.5
     );
-
+    // °F = °C × (9/5) + 32
+    const degFahrenheit = hudData.temperature * (9 / 5) + 32.0;
+    const hudTextE: string = ` Temp: ${hudData.temperature.toFixed(1)}°C / ${degFahrenheit.toFixed(1)}°F`;
+    this.hudComponent.hudBitmap.fillText(
+      hudTextE,
+      this.hudComponent.hudCanvas.width - hudBounds.width / 2,
+      offsetY + tweakParams.lineHeight * 2.5
+    );
+      */
     this.hudComponent.hudBitmap.restore();
   }