+++ src/ts/main.ts
@@ -56,4 +56,5 @@ globalThis.addEventListener("load", () => {
     min: 7.0,
     max: 22.0
   });
+  pane.expanded = false;
 });